chore: move ViewBehaviorOrchestrator to utilities
