+++ packages/web-components/fast-foundation/src/data-grid/data-grid-row.ts
@@ -4,9 +4,9 @@ import {
     FASTElement,
     observable,
     RepeatDirective,
-    ViewBehaviorOrchestrator,
     ViewTemplate,
 } from "@microsoft/fast-element";
+import { ViewBehaviorOrchestrator } from "@microsoft/fast-element/utilities";
 import {
     eventFocusOut,
     eventKeyDown,
